\name{brattle_pal}
\alias{brattle_pal}
\title{Access the Brattle color palette
}
\description{Functions used to access the Brattle color palette by \code{ggplot} functions. \code{BrattleColors} should be preferred; \code{brattle_pal} is used internally by functions in \code{ggplot}.}
\usage{
brattle_pal()
}
\value{Character string of hexidecimal color values
}
\author{
Charlie Gibbons}

\keyword{color}
\keyword{internal}

