\name{BrattleExtras-package}
\alias{BrattleExtras-package}
\alias{BrattleExtras}
\docType{package}
\title{Functions and other objects for Brattle projects}
\description{Functions and other objects for Brattle projects}
\details{
\tabular{ll}{
Package: \tab BrattleExtras\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2017-03-25\cr
License: \tab Licensed only for use by current staff at The Brattle Group\cr
}
}
\author{
Charlie Gibbons

Maintainer: Charlie Gibbons <charlie.gibbons@brattle.com>}
\keyword{ package }
