\name{ImportCalibri}
\alias{ImportCalibri}
\title{Function for importing the Calibri font}
\description{This function imports the Calibri font. The user should not have to call this function directly.}
\usage{
ImportCalibri()
}
\value{The Calibri font is added to the font table from the \code{showtext} package}
\author{
Charlie Gibbons}

\keyword{device}
\keyword{internal}
