\name{scale_color_gradient}
\alias{scale_colour_gradient}
\alias{scale_fill_gradient}
\title{Apply diverging palette to ggplot figures}
\description{These functions wrap the standard ggplot versions to apply a diverging ColorBrewer palette to ggplot figures.}
\usage{
scale_color_gradient2(..., low = "#2166ac", mid = "#f7f7f7", high = "#b2182b")
scale_colour_gradient2(..., low = "#2166ac", mid = "#f7f7f7", high = "#b2182b")
scale_fill_gradient2(..., low = "#2166ac", mid = "#f7f7f7", high = "#b2182b")
}
\arguments{
  \item{\dots}{Arguments passed to corresponding \code{ggplot2} function (see below)}
}
\author{
Charlie Gibbons}
\seealso{\code{\link{ggplot2::scale_colour_gradient2}}, \code{\link{ggplot2::scale_fill_gradient2}} from the \code{ggplot2} package
}

\keyword{color}
